Fix compiler syntax errors caused during merge resolution

+++ SmartRent.API/Services/Implementations/FavoriteService.cs
@@ -183,7 +183,7 @@ namespace SmartRent.API.Services.Implementations
                 },
                 Rating = new PropertyRatingDto
                 {
-                    AverageScore = avgScore,
+                    AverageScore = (decimal)avgScore,
                     TotalReviews = totalReviews
                 }
             };
