+++ SmartRent.API/Services/Implementations/PropertyService.cs
@@ -363,7 +363,7 @@ namespace SmartRent.API.Services.Implementations
                 },
                 Rating = new PropertyRatingDto
                 {
-                    AverageScore = avgScore,
+                    AverageScore = (decimal)avgScore,
                     TotalReviews = totalReviews
                 }
             };
